Changes to quintic task and addition of IMU statistics
Changed quintic task to accept changes to BLE protocol. Added command
getStats which returns statistics for all the NOD modules.

+++ Atmel-SAM4S/BrainMCU_v0.01/BrainMCU_v0.01/src/DebugLog.c
@@ -10,7 +10,7 @@
 #include <string.h>
 
 UINT numberBytes;
-extern unsigned long sgSysTickCount;
+extern uint32_t sgSysTickCount;
 static int size=0;
 static FIL debug_file_object;
 static char DebugFileName[] = "0:DebugLog.txt";
